+++ test/general/__init__.py
@@ -8,6 +8,13 @@ gen_steps = ("generate_early", "create_regions", "create_items", "set_rules", "g
 
 
 def setup_solo_multiworld(world_type: Type[World], steps: Tuple[str, ...] = gen_steps) -> MultiWorld:
+    """
+    Creates a multiworld with a single player of `world_type`, sets default options, and calls provided gen steps.
+    
+    :param world_type: Type of the world to generate a multiworld for
+    :param steps: The gen steps that should be called on the generated multiworld before returning. Default calls
+    steps through pre_fill
+    """
     multiworld = MultiWorld(1)
     multiworld.game[1] = world_type.game
     multiworld.player_name = {1: "Tester"}
