+++ test/general/test_host_yaml.py
@@ -15,6 +15,7 @@ class TestIDs(unittest.TestCase):
             cls.yaml_options = Utils.parse_yaml(f.read())
 
     def test_utils_in_yaml(self) -> None:
+        """Tests that the auto generated host.yaml has default settings in it"""
         for option_key, option_set in Utils.get_default_options().items():
             with self.subTest(option_key):
                 self.assertIn(option_key, self.yaml_options)
@@ -22,6 +23,7 @@ class TestIDs(unittest.TestCase):
                     self.assertIn(sub_option_key, self.yaml_options[option_key])
 
     def test_yaml_in_utils(self) -> None:
+        """Tests that the auto generated host.yaml shows up in reference calls"""
         utils_options = Utils.get_default_options()
         for option_key, option_set in self.yaml_options.items():
             with self.subTest(option_key):
