+++ test/general/test_implemented.py
@@ -5,7 +5,7 @@ from . import setup_solo_multiworld
 
 
 class TestImplemented(unittest.TestCase):
-    def testCompletionCondition(self):
+    def test_completion_condition(self):
         """Ensure a completion condition is set that has requirements."""
         for game_name, world_type in AutoWorldRegister.world_types.items():
             if not world_type.hidden and game_name not in {"Sudoku"}:
@@ -13,7 +13,7 @@ class TestImplemented(unittest.TestCase):
                     multiworld = setup_solo_multiworld(world_type)
                     self.assertFalse(multiworld.completion_condition[1](multiworld.state))
 
-    def testEntranceParents(self):
+    def test_entrance_parents(self):
         """Tests that the parents of created Entrances match the exiting Region."""
         for game_name, world_type in AutoWorldRegister.world_types.items():
             if not world_type.hidden:
@@ -23,7 +23,7 @@ class TestImplemented(unittest.TestCase):
                         for exit in region.exits:
                             self.assertEqual(exit.parent_region, region)
 
-    def testStageMethods(self):
+    def test_stage_methods(self):
         """Tests that worlds don't try to implement certain steps that are only ever called as stage."""
         for game_name, world_type in AutoWorldRegister.world_types.items():
             if not world_type.hidden:
