+++ test/general/test_options.py
@@ -3,7 +3,8 @@ from worlds.AutoWorld import AutoWorldRegister
 
 
 class TestOptions(unittest.TestCase):
-    def testOptionsHaveDocString(self):
+    def test_options_have_doc_string(self):
+        """Test that submitted options have their own specified docstring"""
         for gamename, world_type in AutoWorldRegister.world_types.items():
             if not world_type.hidden:
                 for option_key, option in world_type.options_dataclass.type_hints.items():
