+++ test/netutils/test_location_store.py
